pyproject.toml
[tool.poetry]
name = "aws-lambda-python-streaming"
version = "0.1.0"
description = ""
authors = ["nishide-dev <nishide.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.109.2"
python-dotenv = "0.21.1"
pytest = "^8.0.0"
langchain = "^0.1.9"
langchain-openai = "^0.0.8"
pytest-asyncio = "^0.23.5"
aiosqlite = "^0.20.0"
httpx = "^0.27.0"
uvicorn = "0.23.2"
aws-sam-cli = "^1.111.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
